+++ cohere/client_async.py
@@ -47,6 +47,7 @@ from cohere.responses.chat import AsyncChat, StreamingChat
 from cohere.responses.classify import Example as ClassifyExample
 from cohere.responses.cluster import AsyncClusterJobResult
 from cohere.responses.custom_model import (
+    CUSTOM_MODEL_INTERNAL_STATUS_MAPPING,
     CUSTOM_MODEL_PRODUCT_MAPPING,
     CUSTOM_MODEL_STATUS,
     CUSTOM_MODEL_TYPE,
@@ -1027,7 +1028,7 @@ class AsyncClient(Client):
         """List custom models of your organization.
 
         Args:
-            statuses (CUSTOM_MODEL_STATUS, optional): search for fintunes which are in one of these states
+            statuses (CUSTOM_MODEL_STATUS, optional): search for finetunes which are in one of these states
             before (datetime, optional): search for custom models that were created before this timestamp
             after (datetime, optional): search for custom models that were created after this timestamp
             order_by (Literal["asc", "desc"], optional): sort custom models by created at, either asc or desc
@@ -1038,10 +1039,13 @@ class AsyncClient(Client):
             before = before.replace(tzinfo=before.tzinfo or timezone.utc)
         if after:
             after = after.replace(tzinfo=after.tzinfo or timezone.utc)
-
+        internal_statuses = []
+        if statuses:
+            for status in statuses:
+                internal_statuses.append(CUSTOM_MODEL_INTERNAL_STATUS_MAPPING[status])
         json = {
             "query": {
-                "statuses": statuses,
+                "statuses": internal_statuses,
                 "before": before.isoformat(timespec="seconds") if before else None,
                 "after": after.isoformat(timespec="seconds") if after else None,
                 "orderBy": order_by,
@@ -1187,6 +1191,26 @@ class AsyncClient(Client):
         if not id:
             raise CohereError(message="id must not be empty")
         await self._request(f"{cohere.CONNECTOR_URL}/{id}", method="DELETE")
+
+    async def oauth_authorize_connector(self, id: str, after_token_redirect: str = None) -> str:
+        """Returns a URL which when navigated to will start the OAuth 2.0 flow.
+
+        Args:
+            id (str): The id of your connector
+
+        Returns:
+            str: A URL that starts the OAuth 2.0 flow.
+        """
+        if not id:
+            raise CohereError(message="id must not be empty")
+
+        param_dict = {}
+
+        if after_token_redirect is not None:
+            param_dict["after_token_redirect"] = after_token_redirect
+
+        response = await self._request(f"{cohere.CONNECTOR_URL}/{id}/oauth/authorize", method="GET", params=param_dict)
+        return response["redirect_url"]
 
 
 class AIOHTTPBackend: