+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # Changelog
-## 4.38
+## 4.39
  - [#357] (https://github.com/cohere-ai/cohere-python/pull/357)
     - Embed: add bulk embed support for embed V3
+
+## 4.38
+ - [#351] (https://github.com/cohere-ai/cohere-python/pull/351)
+    - Add connector support
+ - [#356] (https://github.com/cohere-ai/cohere-python/pull/356)
+    - Add connector oauth authorize support
+ - [#345] (https://github.com/cohere-ai/cohere-python/pull/345)
+    - Add mapping for status events for finetunes
 
 ## 4.37
  - [#342] (https://github.com/cohere-ai/cohere-python/pull/342)